SynologyDSMDeviceEntity for disk and volume + add available
